% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compile_proj}
\alias{compile_proj}
\title{Compile project data}
\usage{
compile_proj(search_path, input)
}
\arguments{
\item{input}{}
}
\value{

}
\description{
Compile a dataframe of projects contained within the chosen search directory
}
\keyword{internal}
