% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rename_proj}
\alias{rename_proj}
\title{Rename project files}
\usage{
rename_proj(d, mru, old_name, new_name)
}
\arguments{
\item{d}{Dataframe; All projects found in chosen directory}

\item{old_name}{Character; Name of project to change}

\item{new_name}{Character}
}
\description{
Renames both the project directory and .Rproj file, as well as updating.
}
\details{
Currently changes project_mru as side effect
}
\keyword{internal}
