% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{simplify_mru}
\alias{simplify_mru}
\title{Simplify MRU}
\usage{
simplify_mru(d, mru)
}
\arguments{
\item{mru}{}
}
\value{

}
\description{
creates mru minus projects that appear on RStudio's drop-down menu could be
conoslidated better - or even removed
}
\keyword{internal}
